Implementacion del impresion de datos para Arrendatario
	-Implementacion de impresion : MenuBuscarArrendatarioServiceImpl
	-Implementacion del metodo imprirmiArrendatario : ImpresionServiceImpl
	-Cambio de tipo int a tipo Long en idArrendatario
	-Formateo de todo el programa

+++ src/main/java/com/intelligent/realestate/model/Arrendatario.java
@@ -3,13 +3,13 @@ package com.intelligent.realestate.model;
 public class Arrendatario extends Persona {
 	// TODO: Add comments and the properties that is renting.
 
-	private Integer idArrendatario;
+	private Long idArrendatario;
 
-	public Integer getIdArrendatario() {
+	public Long getIdArrendatario() {
 		return idArrendatario;
 	}
 
-	public void setIdArrendatario(Integer idArrendatario) {
+	public void setIdArrendatario(Long idArrendatario) {
 		this.idArrendatario = idArrendatario;
 	}
